pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta" 

[project]
name = "microseq_tests"
version = "0.0.1"
description= "MicroSeq sandbox package" 
readme= "README.MD"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6",
  "biopython>=1.8",
  "PySide6",
  ]

[tool.setuptools]
package-dir = {"" = "src"}     # tells setuptools hey my code lives here in src/ 

[tools.setuptools.packages.find]
where = ["src"]    # find all packages unders src/

[project.scripts]
microseq = "microseq_tests.microseq:main" 
